made changes based on comments
