+++ aries-site/src/examples/components/toast-notifications/index.js
@@ -1,2 +1,3 @@
 export * from './ToastNotification';
 export * from './ToastNotificationExample';
+export * from './ToastDiagram';
