+++ aries-site/src/pages/components/toast-notifications.mdx
@@ -40,9 +40,7 @@ Due to their limited use cases, toasts are advised to be used sparingly or even 
 
 ## Anatomy 
 
-<Box pad="small">
-  <ToastDiagram />
-</Box>
+<ToastDiagram />
 
 Toasts are compromised of three main elements:
